merge

+++ db/migrate/20131228051822_remove_delivery_manner_id_and_delivery_type_id_and_logistics_company_id_from_order_refund.rb
@@ -1,8 +1,9 @@
 class RemoveDeliveryMannerIdAndDeliveryTypeIdAndLogisticsCompanyIdFromOrderRefund < ActiveRecord::Migration
   def up
-    remove_column :order_refunds, :delivery_manner_id
-    remove_column :order_refunds, :delivery_type_id
-    remove_column :order_refunds, :logistics_company_id
+    remove_column :order_refunds, :delivery_manner_id if column_exists?(:order_refunds, :delivery_manner_id)
+    remove_column :order_refunds, :delivery_type_id if column_exists?(:order_refunds, :delivery_type_id)
+    remove_column :order_refunds, :logistics_company_id if column_exists?(:order_refunds, :logistics_company_id)
+
   end
 
   def down
